refactor : Util 패키지 삭제 및 이름 변경
